changed the type of server ID.
Currently ItemInt is used for the server ID. However, this may
cause problem due to the sign. So their types should be the same.
Now libasura doesn't have ItemUint32. So we have to select 'int'.

+++ test/testFaceRest.cc
@@ -135,7 +135,7 @@ void test_servers(void)
 	for (size_t i = 0; i < NumServerInfo; i++) {
 		g_parser->startElement(i);
 		MonitoringServerInfo &svInfo = serverInfo[i];
-		assertValueInParser(g_parser, "id",        svInfo.id);
+		assertValueInParser(g_parser, "id",   (uint32_t)svInfo.id);
 		assertValueInParser(g_parser, "type", (uint32_t)svInfo.type);
 		assertValueInParser(g_parser, "hostName",  svInfo.hostName);
 		assertValueInParser(g_parser, "ipAddress", svInfo.ipAddress);
